automation-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.0
apscheduler==3.10.4
requests==2.31.0
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
gspread==6.1.0
google-auth==2.29.0
